Lock sub-cli and refactor
