+++ src/passa/cli/lock.py
@@ -2,8 +2,12 @@
 
 from __future__ import absolute_import, print_function, unicode_literals
 
+import argparse
+import os
+
 from resolvelib import NoVersionsAvailable, ResolutionImpossible
 
+from passa.projects import Project
 from passa.reporters import print_requirement
 
 
@@ -26,3 +30,37 @@ def lock(project):
     else:
         success = True
     return success, updated
+
+
+def parse_arguments(argv):
+    parser = argparse.ArgumentParser("passa-lock")
+    parser.add_argument(
+        "--project", dest="project_root",
+        default=os.getcwd(),
+        type=os.path.abspath,
+    )
+    options = parser.parse_args(argv)
+    return options
+
+
+def parsed_main(options):
+    project = Project(options.project_root)
+
+    # Remove the whole lock file to re-lock from scratch.
+    project.lockfile = None
+
+    success, _ = lock(project)
+    if not success:
+        return
+
+    project._l.write()
+    print("Written to project at", project.root)
+
+
+def main(argv=None):
+    options = parse_arguments(argv)
+    parsed_main(options)
+
+
+if __name__ == "__main__":
+    main()
